requirements.txt
aiobotocore==2.0.0
autopep8==1.5.7
flake8==3.9.2
isort==5.9.3
mypy==0.910
pep8==1.7.1
python-dateutil==2.8.2
types-python-dateutil==2.8.2
yapf==0.31.0
